ShuffleReadMetrics + namespacing accumulators
This commit ports ShuffleReadMetrics to use accumulators,
preserving as much of the existing semantics as possible.
It also introduces a nicer way to organize all the internal
accumulators by namespacing them.

+++ core/src/main/scala/org/apache/spark/status/api/v1/api.scala
@@ -178,11 +178,11 @@ class OutputMetrics private[spark](
     val recordsWritten: Long)
 
 class ShuffleReadMetrics private[spark](
-    val remoteBlocksFetched: Int,
-    val localBlocksFetched: Int,
+    val remoteBlocksFetched: Long,
+    val localBlocksFetched: Long,
     val fetchWaitTime: Long,
     val remoteBytesRead: Long,
-    val totalBlocksFetched: Int,
+    val totalBlocksFetched: Long,
     val recordsRead: Long)
 
 class ShuffleWriteMetrics private[spark](
